Add json exchange info files
Add json files for each exchange that contains info specific to that
exchange, such as ticker names used on that exchange, and delimiter used
to define a ticker pair on a market. Also link so that the Exchange class
uses these files as a lookup, so user can enter same ticker symbols for
all exchanges.